Restructured enchant list SQL table (.sql) and randomness (.cpp). Expanded Config file options (.conf.dist).
Changed 'exclusiveSubClass' to 'category' in SQL table as I have removed enchants that require specific weapons, then utilized it to categorize the enchants. (.sql)
Restructured the randomness order of operation to allow for more controlled randomness of enchants. (.cpp)
Added more options in the config file to allow users to tweak the randomness more to their liking. (.conf.dist)